split old test file
